Added 'levels' to GameState, which contain world, npcs, and items for each level ("floor") of the dungeon.

+++ src/util/errors_results.rs
@@ -181,11 +181,11 @@ pub enum DataError {
     MissingNpcDefinition(NpcDefId),
 
     /// Tried to load static world, but no static world defined for id
-    StaticWorldNotFound(u8),
+    StaticWorldNotFound(usize),
 
     /// World needs to fit requirements to be loaded.
     /// * Cannot be larger than `WORLD_WIDTH`x`WORLD_HEIGHT`
-    InvalidWorldFormat(u8),
+    InvalidWorldFormat(usize),
 }
 
 impl fmt::Display for DataError {
